+++ src/onegov.feriennet/onegov/feriennet/sponsors.py
@@ -14,12 +14,14 @@ def load_sponsors(sponsors_path):
 class Sponsor(object):
 
     __slots__ = (
-        'name', 'logo', 'link', 'background', 'height', 'width', 'top'
+        'name', 'logo', 'link', 'background', 'height', 'width', 'top',
+        'banners'
     )
 
     def __init__(self, **kwargs):
         for key, value in kwargs.items():
             setattr(self, key, value)
 
-    def logo_path(self, request):
-        return request.link(StaticFile(self.logo))
+    def url_for(self, request, path):
+        assert path.startswith('sponsors/')
+        return request.link(StaticFile(path))